requirements.txt
streamlit
pandas
unidecode
psycopg2-binary
python-dotenv
plotly